+++ libraries/ts-command-line/src/parameters/CommandLineDefinition.ts
